reference/oap_trust/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "oap-trust"
version = "0.1.0"
description = "OAP reference trust provider — domain and capability attestation for OAP manifests"
readme = "README.md"
license = "CC0-1.0"
requires-python = ">=3.11"
dependencies = [
    "PyJWT[crypto]>=2.8",
    "cryptography>=42.0",
    "dnspython>=2.6",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[project.scripts]
oap-trust = "oap_trust.cli:cli"
oap-trust-api = "oap_trust.api:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
